+++ code/ImageRecognitionApp/ImageRecognitionApp/ViewController.swift
@@ -1,0 +1,25 @@
+//
+//  ViewController.swift
+//  ImageRecognitionApp
+//
+//  Created by Shubham Jindal on 29/03/19.
+//  Copyright © 2019 sjc. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+
+}
+