+++ changelog.md
@@ -1,4 +1,4 @@
 
 # Friday, Week 4, 29/March
-- Initialise the app. [who??]
+- Initialise the app. [Shubham Jindal, jindalsh, 218053165]